requirements.txt
Flask==3.0.3
langchain==0.3.4
langchain-anthropic==0.2.3
langchain-community==0.3.3
langchain-core==0.3.12
langchain-openai==0.2.3
langchain-text-splitters==0.3.0
langgraph==0.2.39
langgraph-checkpoint==2.0.2
langgraph-checkpoint-sqlite==2.0.0
langgraph-sdk==0.1.33
langsmith==0.1.137
numpy==1.26.4
openai==1.52.2
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
python-dotenv==1.0.1
requests==2.32.3
